Sexta actualizacion
Se agregaron nuevas funcionalidades y se mejoro el diseño de la
aplicacion web.

+++ sistema/eliminar_categoria.php
@@ -2,7 +2,13 @@
 	include_once("./conectar.php");
 	$conec=conectar();
 	$id=$_GET["id"];
-	$sql= "DELETE FROM categoria WHERE idCategoria='$id'";
-	mysqli_query($conec,$sql) or die('Error: ' . mysqli_error($con));
-	header("Location: ../listar_categorias.php?error=5");
+	$resul=mysqli_query($conec, "SELECT * FROM subastacategoria WHERE idCategoria='$id'");
+	if(mysqli_num_rows($resul)==0){
+		$sql= "DELETE FROM categoria WHERE idCategoria='$id'";
+		mysqli_query($conec,$sql) or die('Error: ' . mysqli_error($con));
+		header("Location: ../listar_categorias.php?error=5");
+	}
+	else{
+		header("Location: ../listar_categorias.php?error=11");
+	}
 ?>
